bpf: Add generic_start_process_filter for kprobes and uprobes
Kprobes and uprobes have the almost the same code in
generic_kprobe_start_process_filter and
generic_uprobe_start_process_filter it can be generic.
